requirements.txt
psycopg2-binary==2.9.5
cryptocode==0.1
cryptography==37.0.4
Flask==2.2.2
flask-sqlalchemy==3.0.2
Flask-Migrate==4.0.2
Flask-JWT-Extended==4.4.4
